+++ app/models/pwb/import_source.rb
@@ -1,10 +1,23 @@
+# frozen_string_literal: true
+
+# DEPRECATED: This file contains legacy RETS import source definitions.
+# The RETS gem has been removed from the project (Dec 2024).
+# MLS/RETS integration was experimental and never fully implemented.
+# See docs/claude_thoughts/DEPRECATED_FEATURES.md for details.
+#
+# This file is kept for reference but the data is no longer functional.
+
 module Pwb
   # https://github.com/zilkey/active_hash
+  # @deprecated RETS sources are no longer functional - rets gem removed Dec 2024
   class ImportSource < ActiveHash::Base
+    # DEPRECATED: These RETS sources are non-functional.
+    # The rets gem has been removed from the project.
     self.data = [
+      # DEPRECATED - RETS source (non-functional)
       {
         id: 1,
-        source_type: "rets",
+        source_type: "rets",  # DEPRECATED
         unique_name: "mris",
         import_mapper_name: "mls_mris",
         details: {
@@ -16,10 +29,11 @@ module Pwb
         },
         default_property_class: "ALL",
         displayName: "MRIS",
-      # value: "mris"
-      }, {
+      },
+      # DEPRECATED - RETS source (non-functional)
+      {
         id: 2,
-        source_type: "rets",
+        source_type: "rets",  # DEPRECATED
         unique_name: "interealty",
         import_mapper_name: "mls_interealty",
         details: {
@@ -31,7 +45,6 @@ module Pwb
         },
         default_property_class: "PROPERTY",
         displayName: "InterRealty",
-      # value: "interealty"
       },
       # OData source removed - ruby_odata gem no longer supported
     ]